Fix wishlist item 82098, thanks to Ralf Wildenhues:
  ANSIfication of the hp2ps code. The most important changes are the correct
  use of the stdarg mechanism (former hacks could bite on other systems, so
  please tell upstream), inclusion of stdlib.h instead of declaring free
  yourself, adding a few missed PROTO()s and using size_t for xmalloc and
  xrealloc.:


git-svn-id: svn://svn.valgrind.org/valgrind/trunk@2399

+++ massif/hp2ps/Utilities.c
@@ -83,7 +83,7 @@ CommaPrint(fp,n)
 
 void *
 xmalloc(n)
-  int n;
+  size_t n;
 {
     void *r;
 
@@ -98,7 +98,7 @@ xmalloc(n)
 void *
 xrealloc(p, n)
   void *p;
-  int n;
+  size_t n;
 {
     void *r;
     extern void *realloc();
